--- src/Infrastructure/Rule/Exceptions/ArgNotSupportedException.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Exceptions;
-
-use ArtARTs36\MergeRequestLinter\Shared\Exceptions\MergeRequestLinterException;
-
-class ArgNotSupportedException extends MergeRequestLinterException
-{
-    //
-}
