+++ src/Shared/Reflection/TypeResolver/ArrayeeResolver.php
@@ -3,7 +3,6 @@
 namespace ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver;
 
 use ArtARTs36\MergeRequestLinter\Infrastructure\Contracts\Configuration\ArgumentResolver;
-use ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Exceptions\ArgNotSupportedException;
 use ArtARTs36\MergeRequestLinter\Shared\DataStructure\Arrayee;
 use ArtARTs36\MergeRequestLinter\Shared\Reflection\Reflector\Type;
 
@@ -19,7 +18,7 @@ final class ArrayeeResolver implements ArgumentResolver
     public function resolve(Type $type, mixed $value): mixed
     {
         if (! is_array($value)) {
-            throw new ArgNotSupportedException(sprintf(
+            throw new ValueInvalidException(sprintf(
                 'Arg with type %s not supported. Expected type: array',
                 gettype($value),
             ));
