+++ src/Shared/Reflection/TypeResolver/DataObjectResolver.php
@@ -3,7 +3,6 @@
 namespace ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver;
 
 use ArtARTs36\MergeRequestLinter\Infrastructure\Contracts\Configuration\ArgumentResolver;
-use ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Exceptions\ArgNotSupportedException;
 use ArtARTs36\MergeRequestLinter\Shared\Reflection\Reflector\ArrayObjectConverter;
 use ArtARTs36\MergeRequestLinter\Shared\Reflection\Reflector\Type;
 
@@ -27,14 +26,14 @@ final class DataObjectResolver implements ArgumentResolver
         $class = $type->class;
 
         if ($class === null || ! class_exists($class)) {
-            throw new ArgNotSupportedException(sprintf(
+            throw new ValueInvalidException(sprintf(
                 'Type with name "%s" not supported',
                 $class ?? $type->name->value,
             ));
         }
 
         if (! is_array($value) && $value !== null) {
-            throw new ArgNotSupportedException(sprintf(
+            throw new ValueInvalidException(sprintf(
                 'Value for type with name "%s" must be array or null, given %s',
                 $class,
                 gettype($value),
