+++ src/Shared/Reflection/TypeResolver/ObjectCompositeResolver.php
@@ -3,7 +3,6 @@
 namespace ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver;
 
 use ArtARTs36\MergeRequestLinter\Infrastructure\Contracts\Configuration\ArgumentResolver;
-use ArtARTs36\MergeRequestLinter\Infrastructure\Rule\Exceptions\ArgNotSupportedException;
 use ArtARTs36\MergeRequestLinter\Shared\Reflection\Reflector\Type;
 
 final class ObjectCompositeResolver implements ArgumentResolver
@@ -38,7 +37,7 @@ final class ObjectCompositeResolver implements ArgumentResolver
             }
         }
 
-        throw new ArgNotSupportedException(sprintf(
+        throw new ValueInvalidException(sprintf(
             'Resolver for type "%s" not found',
             $type->class,
         ));
