+++ src/Shared/Reflection/TypeResolver/ValueInvalidException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver;
+
+use ArtARTs36\MergeRequestLinter\Shared\Exceptions\MergeRequestLinterException;
+
+final class ValueInvalidException extends MergeRequestLinterException
+{
+    //
+}
